create Invite members to the price reduction procedure 5 day (work)

+++ tender/src/main/java/com/holmech/tender/application/parser/fromexcel/SubjectParseExcel.java
@@ -77,7 +77,7 @@ public class SubjectParseExcel {
                             break;
                         case 2:
                             if (row.getCell(i) != null) {
-                                bufSubject.setPayment(String.valueOf(row.getCell(i).getNumericCellValue()));
+                                bufSubject.setPayment(String.valueOf((int) row.getCell(i).getNumericCellValue()));
                             } else {
                                 bufSubject.setPayment("");
                             }
